mark application tests that rely on external programs as integration tests
